requirements.txt
json
logging
peewee
python-telegram-bot